society-panel/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pyyaml
ray[default]
json_repair
matplotlib
seaborn